move stores to one place
