+++ service/history/workflow/update/store_mock_test.go
@@ -27,22 +27,27 @@ package update_test
 import (
 	"context"
 
+	historypb "go.temporal.io/api/history/v1"
 	"go.temporal.io/api/serviceerror"
 	updatepb "go.temporal.io/api/update/v1"
 
 	persistencespb "go.temporal.io/server/api/persistence/v1"
+	"go.temporal.io/server/internal/effect"
 	"go.temporal.io/server/service/history/workflow/update"
 )
 
-var emptyUpdateStore = mockUpdateStore{
-	VisitUpdatesFunc: func(func(updID string, updInfo *persistencespb.UpdateInfo)) {},
-	GetUpdateOutcomeFunc: func(context.Context, string) (*updatepb.Outcome, error) {
-		return nil, serviceerror.NewNotFound("not found")
-	},
-}
+var (
+	emptyUpdateStore = mockUpdateStore{
+		VisitUpdatesFunc: func(func(updID string, updInfo *persistencespb.UpdateInfo)) {},
+		GetUpdateOutcomeFunc: func(context.Context, string) (*updatepb.Outcome, error) {
+			return nil, serviceerror.NewNotFound("not found")
+		},
+	}
+	eventStoreUnused update.EventStore
+)
 
 type mockUpdateStore struct {
-	update.Store
+	update.UpdateStore
 	VisitUpdatesFunc               func(visitor func(updID string, updInfo *persistencespb.UpdateInfo))
 	GetUpdateOutcomeFunc           func(context.Context, string) (*updatepb.Outcome, error)
 	GetCurrentVersionFunc          func() int64
@@ -76,4 +81,50 @@ func (m mockUpdateStore) IsWorkflowExecutionRunning() bool {
 		return true
 	}
 	return m.IsWorkflowExecutionRunningFunc()
+}
+
+type mockEventStore struct {
+	effect.Controller
+	AddWorkflowExecutionUpdateAcceptedEventFunc func(
+		updateID string,
+		acceptedRequestMessageId string,
+		acceptedRequestSequencingEventId int64,
+		acceptedRequest *updatepb.Request,
+	) (*historypb.HistoryEvent, error)
+
+	AddWorkflowExecutionUpdateCompletedEventFunc func(
+		acceptedEventID int64,
+		resp *updatepb.Response,
+	) (*historypb.HistoryEvent, error)
+
+	CanAddEventFunc func() bool
+}
+
+func (m mockEventStore) AddWorkflowExecutionUpdateAcceptedEvent(
+	updateID string,
+	acceptedRequestMessageId string,
+	acceptedRequestSequencingEventId int64,
+	acceptedRequest *updatepb.Request,
+) (*historypb.HistoryEvent, error) {
+	if m.AddWorkflowExecutionUpdateAcceptedEventFunc != nil {
+		return m.AddWorkflowExecutionUpdateAcceptedEventFunc(updateID, acceptedRequestMessageId, acceptedRequestSequencingEventId, acceptedRequest)
+	}
+	return &historypb.HistoryEvent{EventId: testAcceptedEventID}, nil
+}
+
+func (m mockEventStore) AddWorkflowExecutionUpdateCompletedEvent(
+	acceptedEventID int64,
+	resp *updatepb.Response,
+) (*historypb.HistoryEvent, error) {
+	if m.AddWorkflowExecutionUpdateCompletedEventFunc != nil {
+		return m.AddWorkflowExecutionUpdateCompletedEventFunc(acceptedEventID, resp)
+	}
+	return &historypb.HistoryEvent{}, nil
+}
+
+func (m mockEventStore) CanAddEvent() bool {
+	if m.CanAddEventFunc != nil {
+		return m.CanAddEventFunc()
+	}
+	return true
 }
