+++ service/history/workflow/update/update.go
@@ -31,7 +31,6 @@ import (
 
 	enumspb "go.temporal.io/api/enums/v1"
 	failurepb "go.temporal.io/api/failure/v1"
-	historypb "go.temporal.io/api/history/v1"
 	protocolpb "go.temporal.io/api/protocol/v1"
 	"go.temporal.io/api/serviceerror"
 	updatepb "go.temporal.io/api/update/v1"
@@ -50,35 +49,6 @@ var (
 )
 
 type (
-	// EventStore is the interface that an Update needs to read and write events
-	// and to be notified when buffered writes have been flushed. It is the
-	// expectation of this code that writes to EventStore will return before the
-	// data has been made durable. Callbacks attached to the EventStore via
-	// OnAfterCommit and OnAfterRollback *must* be called after the EventStore
-	// state is successfully written or is discarded.
-	EventStore interface {
-		effect.Controller
-
-		// AddWorkflowExecutionUpdateAcceptedEvent writes an update accepted
-		// event. The data may not be durable when this function returns.
-		AddWorkflowExecutionUpdateAcceptedEvent(
-			updateID string,
-			acceptedRequestMessageId string,
-			acceptedRequestSequencingEventId int64,
-			acceptedRequest *updatepb.Request,
-		) (*historypb.HistoryEvent, error)
-
-		// AddWorkflowExecutionUpdateCompletedEvent writes an update completed
-		// event. The data may not be durable when this function returns.
-		AddWorkflowExecutionUpdateCompletedEvent(
-			acceptedEventID int64,
-			resp *updatepb.Response,
-		) (*historypb.HistoryEvent, error)
-
-		// CanAddEvent returns true if an event can be added to the EventStore.
-		CanAddEvent() bool
-	}
-
 	// Update is a state machine for the update protocol. It reads and writes
 	// messages from the go.temporal.io/api/update/v1 package. See the diagram
 	// in service/history/workflow/update/README.md. The update state machine is
